requirements.txt
flask
telethon
requests
python-dotenv
argostranslate